+++ front/src/applications/operationalStudies/components/ImportTrainSchedule/types.ts
@@ -1,0 +1,5 @@
+export interface Point {
+  longitude: number;
+  latitude: number;
+  name: string;
+}
